% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mod_general_server.R
\name{get_validated_typology}
\alias{get_validated_typology}
\title{Get the last validated typology from reactiveValues}
\usage{
get_validated_typology(validated_selection)
}
\arguments{
\item{validated_selection}{A `reactiveValues` object containing a field
`last_validated_typology` representing the typology last validated by the user.}
}
\value{
A character string indicating the last validated typology (e.g. "france", "region", "departement", "epci", "commune").
}
\description{
This function returns the last validated typology stored in a reactiveValues
object within a Shiny application. It is designed to be called from within the
server logic, by passing the reactiveValues object as an argument.
}
\examples{
# Inside server
typology <- get_validated_typology(validated_selection)
}
