% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_intro_panel.R
\name{mod_intro_panel_fiche_synthese_ui}
\alias{mod_intro_panel_fiche_synthese_ui}
\title{Intro panel for Fiche Synthèse tab}
\usage{
mod_intro_panel_fiche_synthese_ui(id)
}
\arguments{
\item{id}{Module ID.}
}
\value{
A UI tagList to include in a module.
}
\description{
Intro panel for Fiche Synthèse tab
}
