% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_intro_panel.R
\name{mod_intro_panel_server}
\alias{mod_intro_panel_server}
\title{Intro panel server logic with dynamic download}
\usage{
mod_intro_panel_server(
  id,
  thematique,
  validated_selection,
  data_cat_terr,
  data_object,
  download_filename = "data_export"
)
}
\arguments{
\item{id}{Module ID}

\item{thematique}{Name of the thematic tab}

\item{validated_selection}{Reactive input used for generating the intro}

\item{data_cat_terr}{Data passed to `generate_text_intro`}

\item{data_object}{A data.frame to download (e.g., accueil_velo_epci)}

\item{download_filename}{Name for the downloaded file (e.g., "accueil_velo_data")}
}
\description{
Intro panel server logic with dynamic download
}
