% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_intro_panel.R
\name{mod_intro_panel_server_fiche_synthese}
\alias{mod_intro_panel_server_fiche_synthese}
\title{Intro panel server logic for Fiche Synthèse tab}
\usage{
mod_intro_panel_server_fiche_synthese(
  id,
  validated_selection,
  data_cat_terr,
  admin_sf = NULL
)
}
\arguments{
\item{id}{Module ID}

\item{validated_selection}{Reactive input used for generating the intro}

\item{data_cat_terr}{Data passed to `generate_text_intro`}

\item{admin_sf}{Administrative data, containing population and surface data}
}
\description{
Intro panel server logic for Fiche Synthèse tab
}
