% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_intro_panel.R
\name{mod_intro_panel_ui}
\alias{mod_intro_panel_ui}
\title{Intro panel}
\usage{
mod_intro_panel_ui(id, thematique)
}
\arguments{
\item{id}{Module ID.}

\item{thematique}{Name of the thematic tab}
}
\value{
A UI tagList to include in a module.
}
\description{
Intro panel
}
